more cleanups to clarify prototype
